Changes to support windows service

+++ include/dns/dnscommon.hpp
@@ -126,8 +126,8 @@ enum PacketType {
 
 struct BytePacketBuffer {
   std::array<uint8_t, MAX_PACKET_SZ> buf{};
-  uint16_t pos{0};
-  uint16_t size{0};
+  size_t pos{0};
+  size_t size{0};
 };
 
 class DnsObject {
